Generate structs from golang/tools internal/lsp/protocol
It would be best to import this like a normal package, but
gopls team doesn't feel confident about the stability of their
generator and so this remains under internal for now.

That said this is still currently the most accurate and up to date
structs we can get for Go today with little effort.